Move Loki editor to components and rename
